Organize Python code into modules
