+++ src/kolibri_android/android_utils.py
@@ -24,12 +24,15 @@ NotificationManager = autoclass("android.app.NotificationManager")
 PackageManager = autoclass("android.content.pm.PackageManager")
 PendingIntent = autoclass("android.app.PendingIntent")
 PythonActivity = autoclass("org.kivy.android.PythonActivity")
+Secure = autoclass("android.provider.Settings$Secure")
 Timezone = autoclass("java.util.TimeZone")
 Uri = autoclass("android.net.Uri")
 
 ANDROID_VERSION = autoclass("android.os.Build$VERSION")
 RELEASE = ANDROID_VERSION.RELEASE
 SDK_INT = ANDROID_VERSION.SDK_INT
+
+BAD_ANDROID_IDS = ("9774d56d682e549c",)
 
 
 # Path.is_relative_to only on python 3.9+.
@@ -65,6 +68,19 @@ def start_service(service_name, service_args=None):
     service_args = service_args or {}
     service = autoclass("org.endlessos.Key.Service{}".format(service_name.title()))
     service.start(PythonActivity.mActivity, json.dumps(dict(service_args)))
+
+
+def get_android_id():
+    android_id = Secure.getString(
+        get_activity().getContentResolver(), Secure.ANDROID_ID
+    )
+
+    # Don't use this if the retrieved id is falsy, too short, or a specific
+    # id that is known to be hardcoded in many devices.
+    if android_id is None or len(android_id) < 16 or android_id in BAD_ANDROID_IDS:
+        return None
+
+    return android_id
 
 
 def get_service_args():
