+++ src/kolibri_android/kolibri_settings.py
