+++ src/kolibri_android/workers/__main__.py
@@ -1,0 +1,13 @@
+from ..globals import initialize
+
+initialize()
+
+
+def main():
+    from .application import Application
+
+    application = Application()
+    application.run()
+
+
+main()
